--- src/components/Header.tsx
@@ -1,76 +1,0 @@
-import { matchPath, useLocation, useNavigate } from "react-router-dom";
-import { BackIcon, Location, Logo, SearchBlue } from "../assets/svg";
-import { PATH_NAMES } from "../constants/nav-menu";
-
-const Header = () => {
-  const location = useLocation();
-
-  const getPageName = (pathname: string): string | undefined => {
-    for (const [path, name] of PATH_NAMES) {
-      if (matchPath(path, pathname)) {
-        return name;
-      }
-    }
-    return undefined;
-  };
-
-  const isMainHeader = (): boolean =>
-    ["/", "/fitness", "/fitness/:id", "/upload-review/:id", "/my", "/my/pay-history", "/update-review/:id"].some((path) =>
-      matchPath(path, location.pathname)
-    );
-
-  const pageName = getPageName(location.pathname);
-
-  if (!pageName && !isMainHeader()) {
-    return null;
-  }
-
-  return (
-    <div className="w-full max-w-content top-0 h-header bg-white-100 fixed flex items-end z-40">
-      {isMainHeader() ? <MainHeader /> : <PageHeader pageName={pageName} />}
-    </div>
-  );
-};
-
-const MainHeader = () => {
-  const navigate = useNavigate();
-
-  return (
-    <div className="w-full flex justify-center gap-6 border-b-2 border-gray-300 pb-4">
-      <Logo width={"47px"} onClick={() => navigate("/")} />
-      <div className="w-[219px] h-10 bg-gray-200 flex gap-2 rounded-5 px-3 py-2">
-        <Location width={"16px"} />
-        <input
-          type="text"
-          className="w-full bg-transparent text-13px text-gray-500"
-          placeholder={localStorage.getItem("address_name") || "중구 필동로 123번지"}
-          onClick={() => navigate("/set-location")}
-        />
-      </div>
-      <SearchBlue
-        width={"26px"}
-        onClick={() => navigate("/search-fitness")}
-        className="cursor-pointer"
-      />
-    </div>
-  );
-};
-
-interface PageHeaderProps {
-  pageName?: string;
-}
-
-const PageHeader = ({ pageName }: PageHeaderProps) => {
-  return (
-    <div className="w-full h-[27px] flex items-center justify-center relative border-b-2 border-gray-300 pb-8">
-      <BackIcon
-        width={"11px"}
-        className="absolute left-0 cursor-pointer ml-3"
-        onClick={() => window.history.back()}
-      />
-      <p className="text-18px text-black-700 font-bold">{pageName}</p>
-    </div>
-  );
-};
-
-export default Header;
